+++ CHANGELOG.md
@@ -1,5 +1,34 @@
 # Changelog
 
+## [15.12.0](https://github.com/equinor/terraform-azurerm-web-app/compare/v15.11.0...v15.12.0) (2025-02-04)
+
+
+### Features
+
+* configure 32-bit worker ([#232](https://github.com/equinor/terraform-azurerm-web-app/issues/232)) ([8341989](https://github.com/equinor/terraform-azurerm-web-app/commit/8341989a50d757a099b1da65c7d250dfccf38426))
+
+## [15.11.0](https://github.com/equinor/terraform-azurerm-web-app/compare/v15.10.2...v15.11.0) (2025-02-04)
+
+
+### Features
+
+* configure basic authentication for publishing ([#230](https://github.com/equinor/terraform-azurerm-web-app/issues/230)) ([0381a64](https://github.com/equinor/terraform-azurerm-web-app/commit/0381a641b7203a987b7fb9b7bfaf426868404d55))
+* configure HTTP/2 protocol ([#229](https://github.com/equinor/terraform-azurerm-web-app/issues/229)) ([f0aff3f](https://github.com/equinor/terraform-azurerm-web-app/commit/f0aff3f9d55150bf31b62b04ff57be58261eebd4))
+
+## [15.10.2](https://github.com/equinor/terraform-azurerm-web-app/compare/v15.10.1...v15.10.2) (2025-01-30)
+
+
+### Documentation
+
+* reformat notes ([#227](https://github.com/equinor/terraform-azurerm-web-app/issues/227)) ([383c00d](https://github.com/equinor/terraform-azurerm-web-app/commit/383c00d398d4f0d914601285b380aaad0146521b))
+
+## [15.10.1](https://github.com/equinor/terraform-azurerm-web-app/compare/v15.10.0...v15.10.1) (2025-01-30)
+
+
+### Documentation
+
+* update readme ([#225](https://github.com/equinor/terraform-azurerm-web-app/issues/225)) ([fdcd61d](https://github.com/equinor/terraform-azurerm-web-app/commit/fdcd61d46008545af69bc8b42a9fc40d2b422103))
+
 ## [15.10.0](https://github.com/equinor/terraform-azurerm-web-app/compare/v15.9.0...v15.10.0) (2024-11-28)